Add function event to uiEvent_t

+++ firmware/include/functions/fw_codeplug.h
@@ -139,6 +139,6 @@ int codeplugContactGetFreeIndex(void);
 bool codeplugContactGetRXGroup(int index);
 void codeplugInitChannelsPerZone(void);
 bool codeplugGetOpenGD77CustomData(codeplugCustomDataType_t dataType,uint8_t *dataBuf);
-int codeplugGetQuickkeyFunctionID(int key);
+uint16_t codeplugGetQuickkeyFunctionID(int key);
 
 #endif
